api/UploadBill/requirements.txt
azure-functions>=1.12.0
azure-storage-blob>=12.8.0
